Cluster labels (#69)
* committing changes for cluster-lables and fixing the acm-mch install to
set `disableHubSelfManagement' to boolean true instead of string true

* updating managed hub labels template to look for the self-managed label

* adding policy namespace to cluster-labels/values.yaml

* creating new policy namespace

* updating policy_namespace for managed hub

* updating NS creation definition with a label indicating that the NS was
specifically created/would be created by autoshift (detect autoshift
policy namespace collisions in hubofhubs model)

* updating cluster-labels to use a single template and label
reconciliation codeblock
also updating to attempt to pull the label prefix from the config maps
in the cluster, then defaulting to a values object provided one then
defaulting to a hardcoded one if none of the previous sources provided a
label prefix

* removing deepCopy calls inside of the ACM tepmlating section since that
sprig function is not made available in the ACM tepmlating environment

* updating global values to ensure cluster labels override conflicting
cluster set labels

* lookup with label selector appears to have issues so moving to indexing
into the namespace definition and then pulling the label out of there

* updating label application to attempt to pull the policy namespace for a
target cluster from its labels. this is necessary for a managed hub
cluster so that it can pull config maps from the right namespace when
that differs from the policy namespace that the helm chart is aware of.
might be able to do a namespace neutral config map lookup but that
design decision needs to be reviewed

* adding in gitops to hub1 and hub2

* removing acs from the hub of hubs

---------

Co-authored-by: John Harmon <placeholder>